+++ src/main/java/duke/Duke.java
@@ -1,9 +1,15 @@
 package duke;
 
+import duke.command.Command;
+import duke.exceptions.DukeException;
+import duke.parser.Parser;
+import duke.storage.Storage;
 import duke.task.Deadline;
 import duke.task.Event;
 import duke.task.Task;
+import duke.task.TaskList;
 import duke.task.Todo;
+import duke.ui.Ui;
 
 import java.util.Scanner;
