+++ src/main/java/duke/command/Command.java
@@ -1,4 +1,4 @@
-package duke;
+package duke.command;
 
 public enum Command {
     BYE, LIST, MARK, UNMARK, DELETE, TODO, EVENT, DEADLINE
